pyproject.toml
[tool.poetry]
name = "did-wba-server"
version = "0.1.0"
description = "DID WBA Server Example"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "api" },
    { include = "auth" },
    { include = "core" },
    { include = "utils" }
]

[tool.poetry.dependencies]
pyyaml = "^6.0.2"
python = ">=3.10,<4.0"
fastapi = "^0.103.0"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"
agent-connect = "^0.3.5"
python-multipart = "^0.0.6"
aiohttp = "^3.8.5"
cryptography = "^43.0.3"
canonicaljson = "^2.0.0"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.0"
httpx = "^0.28.1"
mcp = {extras = ["cli"], version = ">=1.6.0"}
openai = "^1.77.0"
jinja2 = "^3.1.6"
loguru = "^0.7.3"
pycryptodome = "^3.22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[tool.uv.workspace]
members = ["mcp_client_example"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
